+++ lib/screens/cart_screen/controllers/cart_item_controller.dart
@@ -1,5 +1,6 @@
 import 'package:darkak_e_commerce_app/screens/cart_screen/models/cart_item_model.dart';
 import 'package:darkak_e_commerce_app/services/api_service.dart';
+import 'package:darkak_e_commerce_app/services/connectivity_service.dart';
 import 'package:darkak_e_commerce_app/services/shared_preferences_service.dart';
 import 'package:darkak_e_commerce_app/utils/app_urls.dart';
 import 'package:darkak_e_commerce_app/widgets/styles.dart';
@@ -10,6 +11,11 @@ class CartItemController extends GetxController {
   List<CartItemModel> cartItemsData = [];
 
   Future<void> getCartItem() async {
+    if (!await ConnectivityService.isConnected()) {
+      customErrorMessage(
+          message: 'Please check your internet connection');
+      return;
+    }
    _setLoading(true);
     try {
       final response = await ApiService().get(url: AppUrls.getCartItemUrl, headers: AppUrls.getHeaderWithToken);
@@ -35,6 +41,11 @@ class CartItemController extends GetxController {
   }
 
   Future<void> removeCartItem(String id)async{
+    if (!await ConnectivityService.isConnected()) {
+      customErrorMessage(
+          message: 'Please check your internet connection');
+      return;
+    }
     try {
       final response = await ApiService().delete(url: '${AppUrls.removeCartItemUrl}/$id', headers: AppUrls.getHeaderWithToken);
       if (response.success) {
