+++ lib/screens/wishlist_screen/controllers/wishlist_item_controller.dart
@@ -1,5 +1,6 @@
 import 'package:darkak_e_commerce_app/screens/shop_screen/models/product_model.dart';
 import 'package:darkak_e_commerce_app/services/api_service.dart';
+import 'package:darkak_e_commerce_app/services/connectivity_service.dart';
 import 'package:darkak_e_commerce_app/services/shared_preferences_service.dart';
 import 'package:darkak_e_commerce_app/utils/app_urls.dart';
 import 'package:darkak_e_commerce_app/widgets/styles.dart';
@@ -10,6 +11,11 @@ class WishListItemController extends GetxController {
   List<ProductModel> wishListItemsData = [];
 
   Future<void> getWishListItem() async {
+    if (!await ConnectivityService.isConnected()) {
+      customErrorMessage(
+          message: 'Please check your internet connection');
+      return;
+    }
     _setLoading(true);
     try {
       final response = await ApiService().get(
@@ -38,6 +44,11 @@ class WishListItemController extends GetxController {
   }
 
   Future<void> removeWishListItem(String id) async {
+    if (!await ConnectivityService.isConnected()) {
+      customErrorMessage(
+          message: 'Please check your internet connection');
+      return;
+    }
     try {
       final response = await ApiService().delete(
           url: '${AppUrls.removeWishListItemUrl}/$id',
